+++ src/Service/S3/src/Input/GetObjectRequest.php
@@ -351,13 +351,13 @@ final class GetObjectRequest extends Input
             $headers['If-Match'] = $this->ifMatch;
         }
         if (null !== $this->ifModifiedSince) {
-            $headers['If-Modified-Since'] = $this->ifModifiedSince->format(\DateTimeInterface::RFC822);
+            $headers['If-Modified-Since'] = $this->ifModifiedSince->setTimezone(new \DateTimeZone('GMT'))->format(\DateTimeInterface::RFC7231);
         }
         if (null !== $this->ifNoneMatch) {
             $headers['If-None-Match'] = $this->ifNoneMatch;
         }
         if (null !== $this->ifUnmodifiedSince) {
-            $headers['If-Unmodified-Since'] = $this->ifUnmodifiedSince->format(\DateTimeInterface::RFC822);
+            $headers['If-Unmodified-Since'] = $this->ifUnmodifiedSince->setTimezone(new \DateTimeZone('GMT'))->format(\DateTimeInterface::RFC7231);
         }
         if (null !== $this->range) {
             $headers['Range'] = $this->range;
@@ -405,7 +405,7 @@ final class GetObjectRequest extends Input
             $query['response-content-type'] = $this->responseContentType;
         }
         if (null !== $this->responseExpires) {
-            $query['response-expires'] = $this->responseExpires->format(\DateTimeInterface::RFC822);
+            $query['response-expires'] = $this->responseExpires->setTimezone(new \DateTimeZone('GMT'))->format(\DateTimeInterface::RFC7231);
         }
         if (null !== $this->versionId) {
             $query['versionId'] = $this->versionId;