+++ src/Service/S3/src/Input/PutObjectRequest.php
@@ -701,7 +701,7 @@ final class PutObjectRequest extends Input
             $headers['x-amz-checksum-sha256'] = $this->checksumSha256;
         }
         if (null !== $this->expires) {
-            $headers['Expires'] = $this->expires->format(\DateTimeInterface::RFC822);
+            $headers['Expires'] = $this->expires->setTimezone(new \DateTimeZone('GMT'))->format(\DateTimeInterface::RFC7231);
         }
         if (null !== $this->grantFullControl) {
             $headers['x-amz-grant-full-control'] = $this->grantFullControl;